refactor: collapse chunk runtime and remove pipeline chunker
