+++ internal/parser/chunk/helpers.go
@@ -1,0 +1,43 @@
+//
+// Copyright 2026 The InfiniFlow Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+package chunk
+
+import "unicode"
+
+// DetectLanguage returns a best-effort language code based on the
+// proportion of CJK characters.
+func DetectLanguage(text string) string {
+	cjk := 0
+	total := 0
+	for _, r := range text {
+		if unicode.Is(unicode.Han, r) {
+			cjk++
+		}
+		if unicode.IsLetter(r) {
+			total++
+		}
+	}
+	if total > 0 && float64(cjk)/float64(total) > 0.3 {
+		return "zh"
+	}
+	return "en"
+}
+
+// RuneCount returns the number of runes in text.
+func RuneCount(text string) int {
+	return len([]rune(text))
+}
